pyproject.toml
[project]
name = "stock-predictor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=4.2,<6.0",
    "djangorestframework>=3.16.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "yfinance>=0.2.66",
]
